src/third_party/blink/web_tests/platform/mac/virtual/threaded/external/wpt/css/css-scroll-snap/scroll-margin-expected.txt

This is a testharness.js-based test.
PASS Snaps to the positions adjusted by scroll-margin
FAIL scroll-margin doesn't contribute to the snap position of the element if it's outside of the scroll port assert_equals: expected 0 but got 400
Harness: the test ran to completion.
